% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{methods}
\name{asServerObject,ANY,JuliaObject-method}
\alias{asServerObject,ANY,JuliaObject-method}
\title{Default Julia method for asServerObject()}
\usage{
\S4method{asServerObject}{ANY,JuliaObject}(object, prototype)
}
\description{
the default method for JuliaObject is modelled on the overall default method in XR
}

