% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{methods}
\name{asServerObject,array,JuliaObject-method}
\alias{asServerObject,array,JuliaObject-method}
\title{Array method for asServerObject()}
\usage{
\S4method{asServerObject}{array,JuliaObject}(object, prototype)
}
\description{
use the reshape() function in Julia to create a suitable multi-way array
}

