% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{methods}
\name{asServerObject,list,JuliaObject-method}
\alias{asServerObject,list,JuliaObject-method}
\title{List method for asServerObject()}
\usage{
\S4method{asServerObject}{list,JuliaObject}(object, prototype)
}
\description{
For "plain" lists, produces the Julia expression for a list or a dictionary;
with other attributes, uses the .RClass form.
}

