% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{findJulia}
\alias{findJulia}
\title{Find a Julia Executable}
\usage{
findJulia(test = FALSE)
}
\arguments{
\item{test}{Should the function test for the existence of the application.  Default \code{FALSE}. Calling with
\code{TRUE} is useful to bullet-proof examples or tests for the absence of Julia. If the test search succeeds,
the location is saved in environment variable \code{JULIA_BIN}.}
}
\value{
The location as a character string, unless \code{test} is \code{TRUE}, in which case success or failure
is returned, and the location found (or the empty string) is saved as the environment varialbe.
If \code{test} is \code{FALSE}, failure to find a Julia
in the current system is an error.
}
\description{
This function looks for an executable Julia application in the local operating system.  The location can be prespecified by
setting environment variable \code{JULIA_BIN}; otherwise, the function looks in various conventional locations
and if that doesn't work, runs a shell command to look for \code{julia}.
}

