% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{class}
\name{from_Julia-class}
\alias{from_Julia-class}
\title{Class for General Julia Composite Type Objects}
\description{
The Julia side of the interface will return a general object from a composite type as an R
object of this class.  Its Julia fields (converted to R objects) can be accessed by the \code{$}
operator.
}
\section{Slots}{

\describe{
\item{\code{Class}}{the Julia type.}

\item{\code{data}}{the converted Julia data, as a list whose names are the names of the Julia fields.}
}}

