% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{class}
\name{JuliaFunction-class}
\alias{JuliaFunction}
\alias{JuliaFunction-class}
\alias{initialize,JuliaFunction-method}
\title{Proxy Objects in R for Julia Functions}
\usage{
\S4method{initialize}{JuliaFunction}(.Object, name, module = "",
  evaluator = RJulia(), ...)
}
\description{
A class and generator function for proxies in R for Julia functions.
}
\details{
An object from this class is an R function that is a proxy for a function in Julia. Calls to the R function evaluate
a call to the Julia function.  The arguments in the call are converted to equivalent Julia objects;
these typically include proxy objects for results previously computed through the XRJulia interface.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: creates a function object that calls the named Julia function, optionally imported from
the specified module.
}}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the server language function}

\item{\code{module}}{the name of the module, if that needs to be imported}

\item{\code{evaluatorClass}}{the class for the evaluator, by default and usually, \code{\link{JuliaInteface}}}
}}

