% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{juliaGet}
\alias{juliaGet}
\title{converts the proxy object that is its argument to an \R{} object.}
\usage{
juliaGet(object, evaluator = XR::getInterface(.JuliaInterfaceClass))
}
\description{
converts the proxy object that is its argument to an \R{} object.
}

