% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{juliaImport}
\alias{juliaImport}
\title{Import a Julia module or add a directory to the Julia Search Path}
\usage{
juliaImport(..., evaluator)
}
\arguments{
\item{...}{arguments for the \code{$Import()} method. See the method documentation for details.}
}
\description{
adds the module information specified to the modules imported for Julia evaluators.
}
\details{
If called from the source directory of a package during installation, both \code{juliaImport}
and \code{juliaAddToPath()} also set up
a load action for that package.  The functional versions, not the methods themselves, should
be called from package source files to ensure that the load actions are created.
}

