% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{class}
\name{JuliaObject-class}
\alias{JuliaObject-class}
\alias{JuliaObject}
\title{Proxy Objects in R for Julia Objects}
\description{
This is a class for all proxy objects from a Julia class with an R proxy class definition.
Objects will normally be from a subclass of this class, for the specific Julia class.
}
\details{
Proxy objects returned from the Julia interface will be promoted to objects
from a specific R proxy class for their Julia class, if such a class has been defined.
}

