% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{juliaVersion}
\alias{juliaVersion}
\title{Get or test the Julia Version information}
\usage{
juliaVersion(test = FALSE, .ev = RJulia())
}
\arguments{
\item{test}{If \code{TRUE}, tests compatibility
(currently that the major number is at least \code{1}). Default \code{FALSE}}

\item{.ev}{The evaluator object to use.  By default, and usually, the current evaluator.}
}
\value{
A named list with the
members of the Julia object, the usually relevant ones being \code{"major"}, \code{"minor"} and 
\code{"patch"}. \code{test=TRUE} overrides as described.
}
\description{
The Julia constant structure \code{VERSION} is returned.
 If \code{test} is \code{TRUE}, only returns a logical testing whether
this version is compatible with \code{XRJulia}.
}
