% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{RJulia}
\alias{RJulia}
\title{An Evaluator for the Julia Interface.}
\usage{
RJulia(...)
}
\description{
Returns an evaluator for the Julia interface.  Starts one on the first call, or if arguments are provided;
providing argument \code{.makeNew = TRUE} will force a new evaluator.  Otherwise, the current evaluator is
returned.
}
\details{
See \code{\link{JuliaInterface}} for details of the evaluator.
}

