% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{setJuliaClass}
\alias{setJuliaClass}
\title{Define a Proxy Julia Class (Composite Type)}
\usage{
setJuliaClass(juliaType, module = "", fields = character(),
  where = topenv(parent.frame()), proxyObjectClass = "JuliaObject", ...)
}
\description{
Given the name and optionally the module for a Julia composite type, defines an R proxy
class with the same fields as the Julia type.  By default, uses metadata from Julia to
find the fields.  If the call supplies the desired field names explicitly, metadata is
not used.
}

