\name{XRjulia-package}
\alias{XRjulia-package}
\alias{XRjulia}
\docType{package}
\title{
  Structured Interface to Julia
}
\description{
A Julia interface structured according to the general
form described in package XR and the book (in progress)
"Extending R"
}
\author{
John Chambers

Maintainer: John Chambers <jmc@stat.stanford.edu>

}
\references{
 Chambers, John M. (book in progress)
\emph{Extending R}
}

\keyword{ package }

