% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\docType{class}
\name{from_Julia-class}
\alias{from_Julia-class}
\title{Class for General Julia Composite Type Objects}
\description{
The Julia side of the interface will return a general object from a composite type as an R
object of class "from_Julia.  its Julia fields (converted to R objects) can be accessed by the \code{$}
operator.
}
\section{Slots}{

\describe{
\item{\code{serverClass}}{the Julia type.}

\item{\code{module}}{the Julia module, or ""}

\item{\code{fields}}{the converted versioin of the Julia fields; these are accessed by the \code{$} operator.}
}}

