% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{juliaImport}
\alias{juliaAddToPath}
\alias{juliaImport}
\title{Import a Julia module or add a directory to the Julia Search Path}
\usage{
juliaImport(..., evaluator)

juliaAddToPath(directory = "julia",
  package = utils::packageName(topenv(parent.frame())), pos = NA, evaluator)
}
\arguments{
\item{...}{arguments for the \code{$Import()} method. See the method documentation for details.}

\item{evaluator}{The evaluator object to use. Supplying this argument suppresses the load action.}

\item{directory}{the directory to add, defaults to "julia"}

\item{package,pos}{arguments \code{package} and \code{pos} to the method, usually omitted.}
}
\description{
adds the module information specified to the modules imported for Julia evaluators.
}
\details{
If called from the source directory of a package during installation, both \code{juliaImport}
and \code{juliaAddToPath()} also set up
a load action for that package.  The functional versions, not the methods themselves, should
be called from package source files to ensure that the load actions are created.
}
\section{Functions}{
\itemize{
\item \code{juliaAddToPath}: adds the directory specified to the search path for Julia objects.
}}

