% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{RJulia}
\alias{RJulia}
\title{An Evaluator for the Julia Interface.}
\usage{
RJulia(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{getInterface}()} but none usually required.
See \code{\link{JuliaInterface}} for details of the evaluator.}
}
\description{
Returns an evaluator for the Julia interface.  Starts one on the first call, or if arguments are provided;
providing argument \code{.makeNew = TRUE} will force a new evaluator.  Otherwise, the current evaluator is
returned.
}
