% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJuliaConnect.R
\name{setJuliaClass}
\alias{setJuliaClass}
\title{Define a Proxy Julia Class (Composite Type)}
\usage{
setJuliaClass(juliaType, module = "", fields = character(),
  where = topenv(parent.frame()), proxyObjectClass = "JuliaObject", ...)
}
\arguments{
\item{juliaType, module}{Strings identifying the composite type and optionally the module containing it. In normal use,
metadata from Julia is used to find the definition of the type.}

\item{fields, where, proxyObjectClass, ...}{Overriding arguments that should not be used by direct calls from package source code.}
}
\description{
Given the name and optionally the module for a Julia composite type, defines an R proxy
class with the same fields as the Julia type.  By default, uses metadata from Julia to
find the fields.  If the call supplies the desired field names explicitly, metadata is
not used.
}
